ActionDto  equality operator overriding is updated to the elegant form it can be
Some people sees manually overriding the equality operator (==) is an overhead or it may be a boilerplate code (if all models require equality operator to be overridden).  so if you have many models that should override that method, I suggest using the Equatable package. just was passing from here. My Greetings